resource_hydra_project: improve error handling for unexpected responses
Now, if we ever experience an unexpected response, we print out what the server
returned (e.g. 500 Internal Server Error) and the responding JSON. This makes
the error handling more robust in that we no longer need to check every possible
response; as long as we get the expected response, everything is (assumed to be)
A-okay.