hydra_project: add a test case of a hidden and disabled project
```
=== RUN   TestAccHydraProject_hiddenDisabled
    resource_hydra_project_test.go:62: Step 1/1 error: After applying this test step and performing a `terraform refresh`, the plan was not empty.
        stdout

        An execution plan has been generated and is shown below.
        Resource actions are indicated with the following symbols:
          ~ update in-place

        Terraform will perform the following actions:

          # hydra_project.test will be updated in-place
          ~ resource "hydra_project" "test" {
              ~ enabled      = true -> false
                id           = "p2vogpae"
                name         = "p2vogpae"
              ~ visible      = true -> false
                # (4 unchanged attributes hidden)
            }

        Plan: 0 to add, 1 to change, 0 to destroy.
--- FAIL: TestAccHydraProject_hiddenDisabled (6.23s)
```

This is because of https://github.com/NixOS/hydra/blob/master/src/lib/Hydra/Controller/Project.pm#L157-L158:

        , enabled => defined $c->stash->{params}->{enabled} ? 1 : 0
        , hidden => defined $c->stash->{params}->{visible} ? 0 : 1

which means if the parameter is provided *at all* it is assumed
to be `true`.

One fix would be for us to not send the field if we want it to be
false. Another, more expensive but nicer fix would be to make hydra's
code use `defined` and also see if the value is also truthy.