wip: make the tests pass
1. I noticed you were using the same props over and over, which I think
is not good since it'll reuse that same memory over and over. So, I make a new props in the loop.

2. I added this line in the loop over keys:

    log.Printf("[DEBUG] %s", k)

and wouldn't you know it:

    [DEBUG]
    [DEBUG]
    [DEBUG] nixpkgs
    [DEBUG] ofborg

What is going on here?

    keys := make([]string, len(in)

says give me a list pre-filled len(in) long of nulls. Changing this to:

    keys := make([]string, 0, len(in))

made it allocate len(in) worth of memory, but not pre-fill the slice.
This explains this error:

    - .input: planned set element cty.ObjectVal(map[string]cty.Value{"name":cty.StringVal("nixpkgs"), "notify_committers":cty.False, "type":cty.StringVal("git"), "value":cty.StringVal("https://github.com/NixOS/nixpkgs.git nixpkgs-unstable")}) does not correlate with any element in actual
    - .input: planned set element cty.ObjectVal(map[string]cty.Value{"name":cty.StringVal("ofborg"), "notify_committers":cty.False, "type":cty.StringVal("git"), "value":cty.StringVal("https://github.com/nixos/ofborg.git released")}) does not correlate with any element in actual

How?

    out := make([]interface{}, 0, len(in))

was returning a list of:

    [ null, null, nixpkgs, ofborg ]

thus ... it was getting 2 more inputs than expected, and not yet even
realizing the first two were null.

...

clean-up of this patch left as an exercise :).

PS: Consider adding more logs, all over the place :).