Split out examples
We now have two example setups: `declarative`, which will set up jobsets based
upon the provided `declarative` block (disabled by default to prevent evaluating
and building Nixpkgs and NixOS Small); and `default`, which will set up a
Nixpkgs project and two jobsets -- both for `trunk`, but one is a legacy type
and the other is a flake type (`check_interval` is set to 0 to prevent
evaluating and building Nixpkgs).

This split is necessary so we don't accidentally exemplify creating a
declarative project with Terraform-managed jobsets (the Hydra web page doesn't
allow creating jobsets for declarative projects, so at the very least, we
shouldn't make it easy to accidentally do so, in the case that Hydra doesn't
barf when doing this via the HTTP API).